Zmiana timera na nieskończoną pętlę. Wyświetlanie FPS. Początek obsługi funkcjonalności pauzy. Naprawiony błąd z floatami w klawiaturze. Zaktualizowane schematy.

+++ klawiatura.h
@@ -16,7 +16,7 @@ class Klawiatura : public QObject, public UrzadzenieWejscia
 		int nawigator;
 		int nawigatorPolozenie;
 		int nawigatorWcisniete;
-		float dzojstiki[2];
+		int dzojstiki[2];
 
 	private:
 		Przycisk keyNaPrzycisk(int key);
